Unlocked the secret of the raio. For CSS borders, there can be two values, the horizontal and vertical. If the vertical is not specified, the value of the horizontal is used. This is fine in the example I found of a 100x50 box with corner radius of 75px - everything was fixed - but when I was using percentages, that didn't translate straight to 75% because that's only the percent of the width (100). Height needed to be 150.
Played a bit with the border too, and using that for any part is going
to be tricky because borders must be a fixed value, which means I'll
likely need to do two circles instead, one for the border and one for
the body.

Next I am going to play with the height (and by extension the radius
values again) because a 50% height doesn't do it for me. I need it maybe
around 90? Maybe not, but that's what I'm going to do now.