Shit, that actually works pretty well.
A couple of things, again. The ratio (75% of w, 150% of h) is reliable
and consistant for making a half circle, or at least something that
looks half-circle enough for me.

Two divs, one for the border and one for the body also works. It's
likely not even necessary to add numbers, the imagery is so solid that
it shouldn't be a problem, though maybe I want them anyway. The nice
thing about no numbers is that in theory at least the 9 and 3 should
only be half visible, but if I do that I'm going to force some
scrolling. If I just through some divs up for the hour markers and
figure out where they need to be correctly, people will figure it out
and the message will be conveyed just fine.